+++ src/App.js
@@ -20,6 +20,7 @@ import SalesHistoryPage from './pages/SalesHistoryPage';
 import InventoryReportPage from './pages/InventoryReportPage';
 import DashboardPage from './pages/DashboardPage';
 import ClientesPage from './pages/ClientesPage';
+import CarteraPage from './pages/CarteraPage';
 import ProveedoresPage from './pages/ProveedoresPage';
 import UserManagementPage from './pages/UserManagementPage';
 import AdminCompaniesPage from './pages/AdminCompaniesPage';
@@ -124,6 +125,7 @@ function AppContent() {
           <Route path="/historial-ventas" element={<PrivateRoute><SalesHistoryPage /></PrivateRoute>} />
           <Route path="/reporte-inventario" element={<PrivateRoute requiredRole="admin_compania"><InventoryReportPage /></PrivateRoute>} />
           <Route path="/clientes" element={<PrivateRoute><ClientesPage /></PrivateRoute>} />
+          <Route path="/cartera" element={<PrivateRoute requiredRole="admin_compania"><CarteraPage /></PrivateRoute>} />
           <Route path="/proveedores" element={<PrivateRoute><ProveedoresPage /></PrivateRoute>} />
           <Route path="/dashboard" element={<PrivateRoute><DashboardPage /></PrivateRoute>} />
           <Route path="/gestion-usuarios" element={<PrivateRoute requiredRole="admin_compania"><UserManagementPage /></PrivateRoute>} />